genai-usecases/conversational-analytics/backend/requirements.txt
fastapi
uvicorn
pymongo
pydantic
pyjwt
passlib[bcrypt]

python-multipart
python-dotenv

langchain-core
langchain-groq
langchain-google-genai
